Conflicts:
	ABACUS.1.0.0/source/src_global/lapack_connector.h
	ABACUS.1.0.0/source/src_lcao/gint_gamma_rho.cpp
	ABACUS.1.0.0/source/src_lcao/gint_gamma_vl.cpp
	ABACUS.1.0.0/source/src_lcao/local_orbital_charge.cpp
	ABACUS.1.0.0/source/src_lcao/local_orbital_elec.cpp
	ABACUS.1.0.0/source/src_pw/charge_mixing.cpp
	ABACUS.1.0.0/source/src_pw/charge_pulay.cpp
	ABACUS.1.0.0/source/src_pw/electrons.cpp
	ABACUS.1.0.0/source/src_pw/unitcell.cpp
	ABACUS.1.0.0/source/src_pw/wavefunc_in_pw.cpp

+++ ABACUS.1.0.0/source/src_external/src_pdiag/pdiag_double.h
@@ -14,7 +14,7 @@ class Pdiag_Double : public Pdiag_Basic
 
 	int nrow;
 	int ncol;
-	int nloc;
+	long nloc;
 
 #ifdef __MPI
     int blacs_ctxt;
